edited comments
to reflect more of what I did earlier today since I felt a bit rushed and didn't document my work so well on this project. I think this project was good practice of what we learned earlier this week about Redux, but with the added part about thunk/dispatch and the idea of the middleware that intercepts the action creator before it goes to the reducer. I think this project might be too small to really see what is going on with the dispatch when you execute it but it is good to know I can make it work in case I need it for something bigger.